+++ src/pages/Home.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Home() {
+  return (
+    <div>
+      <h1>EasyMoto</h1>
+    </div>
+  )
+}
+
+export default Home
