+++ src/pages/Layout.js
@@ -1,0 +1,25 @@
+import Container from 'react-bootstrap/Container';
+import Nav from 'react-bootstrap/Nav';
+import Navbar from 'react-bootstrap/Navbar';
+import { Outlet } from "react-router-dom";
+
+function Layout() {
+  return (
+    <div>
+      <Navbar bg="dark" variant="dark">
+        <Container>
+          <Navbar.Brand href="/">EasyMoto</Navbar.Brand>
+          <Nav className="me-auto">
+            <Nav.Link href="/">Home</Nav.Link>
+            <Nav.Link href="./produtos">Produtos</Nav.Link>
+            <Nav.Link href="utilizadores">Utilizadores</Nav.Link>
+            <Nav.Link href="categorias">Categorias</Nav.Link>
+          </Nav>
+        </Container>
+      </Navbar>
+      <Outlet />
+    </div>
+  )
+}
+
+export default Layout
